Life_ocr1/requirement_life_ocr.txt
Django
djangorestframework
PyMuPDF # for `fitz`
pytesseract
pdf2image
PyPDF2
pdfplumber
numpy
opencv-python
requests
google-generativeai
Pillow
dotenv